requirements.txt
discord
python-dotenv
pyngrok